Remove Applicative/Monad instances for Handle
This is a bit painful, but it fixes #13, and it doesn’t actually cause
any problems for the effect handlers currently provided by eff. The
root issue is that `locally` doesn’t make sense unless it’s the only
operation taken by a handler, since the handler may not be on the stack
by the time it returns.

I would love to provide a more principled treatment of scoped
operations, but recent investigation suggests there are some fundamental
tensions that are difficult to resolve. I’m still exploring that, so I
don’t have anything conclusive to say yet, but this workaround seems
sound for now.